requirements.txt
# RAG 知识库问答系统 v2.0 依赖

# LangChain 核心
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20

# 向量数据库
chromadb>=0.4.0

# 检索增强
rank-bm25>=0.2.2

# OpenAI
openai>=1.0.0

# 环境变量
python-dotenv>=1.0.0

# Token 计算
tiktoken>=0.5.0

# 文档加载
pypdf>=3.0.0
python-docx>=1.0.0
unstructured>=0.10.0

# API 框架
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Gradio UI (面试亮点!)
gradio>=4.0.0

# 其他工具
numpy>=1.24.0
python-multipart>=0.0.6

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
